Moving open_library_information method to books_controller
