+++ config/routes.rb
@@ -7,7 +7,12 @@ Rails.application.routes.draw do
   resources :users, only: %i[index show] do
     resources :rents, only: %i[index create]
   end
-  resources :books, only: %i[index show]
   resources :book_suggestions, only: :create
+  resources :books, only: %i[index show] do
+    collection do
+      get :open_library_information
+    end
+  end
+
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
