--- spec/controllers/book_information_controller_spec.rb
@@ -1,4 +1,0 @@
-require 'rails_helper'
-
-describe BookInformationController do
-end
